Fix segfault with reduced trie for long type introduced in 327476f

+++ src/mkcedar.cc
@@ -18,7 +18,7 @@ int main (int argc, char **argv) {
   if (argc < 3)
     { std::fprintf (stderr, "Usage: %s keys trie\n", argv[0]); std::exit (1); }
   //
-  cedar::da <int> trie;
+  cedar::da <long> trie;
   int n = 0;
   FILE* fp = argv[1][0] == '-' ? stdin : std::fopen (argv[1], "r");
   char line[8192];
@@ -29,7 +29,7 @@ int main (int argc, char **argv) {
   if (trie.save (argv[2]) != 0)
     { std::fprintf (stderr, "cannot save trie: %s\n", argv[2]); std::exit (1); }
   //
-  std::fprintf (stderr, "keys: %ld\n", trie.num_keys ());
+  std::fprintf (stderr, "keys: %zu\n", trie.num_keys ());
   std::fprintf (stderr, "size: %ld\n", trie.size ());
   std::fprintf (stderr, "nonzero_size: %ld\n", trie.nonzero_size ());
   return 0;
